+++ posthog/api/advanced_activity_logs/__init__.py
@@ -1,8 +1,17 @@
 from .field_discovery import AdvancedActivityLogFieldDiscovery
 from .filters import AdvancedActivityLogFilterManager
-from .viewset import AdvancedActivityLogFiltersSerializer, AdvancedActivityLogsViewSet
+from .viewset import (
+    ActivityLogPagination,
+    ActivityLogSerializer,
+    ActivityLogViewSet,
+    AdvancedActivityLogFiltersSerializer,
+    AdvancedActivityLogsViewSet,
+)
 
 __all__ = [
+    "ActivityLogPagination",
+    "ActivityLogSerializer",
+    "ActivityLogViewSet",
     "AdvancedActivityLogsViewSet",
     "AdvancedActivityLogFiltersSerializer",
     "AdvancedActivityLogFilterManager",
