+++ posthog/api/advanced_activity_logs/utils.py
@@ -3,12 +3,10 @@ from typing import Optional
 
 from django.utils import timezone
 
-from posthog.api.advanced_activity_logs.constants import (
-    ADVANCED_ACTIVITY_LOGS_LOOKBACK_FALLBACK_LIMIT,
-    ADVANCED_ACTIVITY_LOGS_LOOKBACK_FALLBACK_UNIT,
-)
 from posthog.constants import AvailableFeature
 from posthog.models import Organization
+
+from .constants import ADVANCED_ACTIVITY_LOGS_LOOKBACK_FALLBACK_LIMIT, ADVANCED_ACTIVITY_LOGS_LOOKBACK_FALLBACK_UNIT
 
 
 def get_activity_log_lookback_restriction(organization: Organization) -> Optional[timezone.datetime]:
@@ -22,7 +20,6 @@ def get_activity_log_lookback_restriction(organization: Organization) -> Optiona
     unit = audit_log_feature.get("unit")
 
     if limit is None or unit is None:
-        # Return fallback values
         limit = ADVANCED_ACTIVITY_LOGS_LOOKBACK_FALLBACK_LIMIT
         unit = ADVANCED_ACTIVITY_LOGS_LOOKBACK_FALLBACK_UNIT
